+++ core/src/main/java/org/mule/runtime/core/processor/strategy/ReactorStreamProcessingStrategyFactory.java
@@ -10,13 +10,12 @@ import static java.lang.Math.min;
 import static java.lang.Runtime.getRuntime;
 import static org.mule.runtime.core.api.processor.ReactiveProcessor.ProcessingType.CPU_LITE;
 import static org.mule.runtime.core.api.processor.ReactiveProcessor.ProcessingType.CPU_LITE_ASYNC;
-import static org.mule.runtime.core.processor.strategy.AbstractRingBufferProcessingStrategy.WaitStrategy.LITE_BLOCKING;
 import static reactor.core.publisher.Flux.from;
 import static reactor.core.scheduler.Schedulers.fromExecutorService;
-import static reactor.util.concurrent.QueueSupplier.SMALL_BUFFER_SIZE;
-import static reactor.util.concurrent.QueueSupplier.isPowerOfTwo;
 
 import org.mule.runtime.api.exception.MuleException;
+import org.mule.runtime.api.lifecycle.Startable;
+import org.mule.runtime.api.lifecycle.Stoppable;
 import org.mule.runtime.api.scheduler.Scheduler;
 import org.mule.runtime.core.api.MuleContext;
 import org.mule.runtime.core.api.processor.ReactiveProcessor;
@@ -35,63 +34,7 @@ import java.util.function.Supplier;
  *
  * @since 4.0
  */
-public class ReactorStreamProcessingStrategyFactory extends AbstractProcessingStrategyFactory {
-
-  protected static String RING_BUFFER_SCHEDULER_NAME_SUFFIX = ".ring-buffer";
-
-  public static final int DEFAULT_BUFFER_SIZE = SMALL_BUFFER_SIZE;
-  public static final int DEFAULT_SUBSCRIBER_COUNT = 1;
-  public static final String DEFAULT_WAIT_STRATEGY = LITE_BLOCKING.name();
-
-  private int bufferSize = DEFAULT_BUFFER_SIZE;
-  private int subscriberCount = DEFAULT_SUBSCRIBER_COUNT;
-  private String waitStrategy = DEFAULT_WAIT_STRATEGY;
-
-  /**
-   * Configure the size of the ring-buffer size used to buffer and de-multiplexes events from multiple source threads. This value
-   * must be a power-of two.
-   * <p/>
-   * Ring buffers typically use a power of two because it means that the rollover at the end of the buffer can be achieved using a
-   * bit mask rather than having to explicitly compare the head/tail pointer with the end of the buffer.
-   *
-   * @param bufferSize buffer size to use.
-   */
-  public void setBufferSize(int bufferSize) {
-    if (!isPowerOfTwo(bufferSize)) {
-      throw new IllegalArgumentException("bufferSize must be a power of 2 : " + bufferSize);
-    }
-    this.bufferSize = bufferSize;
-  }
-
-  /**
-   * Configure the number of ring-buffer subscribers.
-   *
-   * @param subscriberCount
-   */
-  public void setSubscriberCount(int subscriberCount) {
-    this.subscriberCount = subscriberCount;
-  }
-
-  /**
-   * Configure the wait strategy used to wait for new events on ring-buffer.
-   *
-   * @param waitStrategy
-   */
-  public void setWaitStrategy(String waitStrategy) {
-    this.waitStrategy = waitStrategy;
-  }
-
-  protected int getBufferSize() {
-    return bufferSize;
-  }
-
-  protected int getSubscriberCount() {
-    return subscriberCount;
-  }
-
-  protected String getWaitStrategy() {
-    return waitStrategy;
-  }
+public class ReactorStreamProcessingStrategyFactory extends AbstractStreamProcessingStrategyFactory {
 
 
   @Override
@@ -106,7 +49,7 @@ public class ReactorStreamProcessingStrategyFactory extends AbstractProcessingSt
                                                getMaxConcurrency());
   }
 
-  static protected class ReactorStreamProcessingStrategy extends AbstractRingBufferProcessingStrategy {
+  static class ReactorStreamProcessingStrategy extends AbstractStreamProcessingStrategy implements Startable, Stoppable {
 
     private Supplier<Scheduler> cpuLightSchedulerSupplier;
     private Scheduler cpuLightScheduler;
